Return InputStream from azure client
