+++ extensions/azure/storage/AzureDataLakeStorage.cpp
@@ -64,8 +64,8 @@ bool AzureDataLakeStorage::deleteFile(const DeleteAzureDataLakeStorageParameters
 
 std::optional<uint64_t> AzureDataLakeStorage::fetchFile(const FetchAzureDataLakeStorageParameters& params, io::BaseStream& stream) {
   try {
-    AzureDataLakeStorageInputStream input(data_lake_storage_client_->fetchFile(params));
-    return internal::pipe(&input, &stream);
+    auto result = data_lake_storage_client_->fetchFile(params);
+    return internal::pipe(result.get(), &stream);
   } catch (const std::exception& ex) {
     logger_->log_error("An exception occurred while fetching '%s/%s' of filesystem '%s': %s", params.directory_name, params.filename, params.file_system_name, ex.what());
     return std::nullopt;
